--- src/main/java/Application.java
@@ -1,9 +1,0 @@
-import controller.GameController;
-
-public class Application {
-
-    public static void main(String[] args) {
-        GameController gameController = new GameController();
-        gameController.run();
-    }
-}
