+++ src/main/java/Blackjack/Application.java
@@ -1,0 +1,11 @@
+package Blackjack;
+
+import Blackjack.controller.GameController;
+
+public class Application {
+
+    public static void main(String[] args) {
+        GameController gameController = new GameController();
+        gameController.run();
+    }
+}
