+++ src/main/java/Blackjack/domain/card/Card.java
@@ -1,0 +1,8 @@
+package Blackjack.domain.card;
+
+public record Card(Rank rank, Suit suit) {
+
+    public boolean isAce() {
+        return this.rank == Rank.ACE;
+    }
+}
