+++ src/main/java/Blackjack/domain/game/CardDeckGenerator.java
@@ -1,0 +1,25 @@
+package Blackjack.domain.game;
+
+import Blackjack.domain.card.Card;
+import Blackjack.domain.card.Rank;
+import Blackjack.domain.card.Suit;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+
+public class CardDeckGenerator {
+    private static final List<Card> cards = new ArrayList<>();
+
+    static {
+        Arrays.stream(Rank.values())
+                .flatMap(rank -> Arrays.stream(Suit.values())
+                        .map(suit -> new Card(rank, suit)))
+                .forEach(cards::add);
+    }
+
+    public static CardDeck generate() {
+        Collections.shuffle(cards);
+        return new CardDeck(cards);
+    }
+}
