+++ src/main/java/Blackjack/dto/ParticipantCardsDto.java
@@ -1,0 +1,7 @@
+package Blackjack.dto;
+
+import Blackjack.domain.card.Card;
+import java.util.List;
+
+public record ParticipantCardsDto(String name, List<Card> cards, int cardsScore) {
+}
