+++ src/main/java/Blackjack/exception/ErrorException.java
@@ -1,0 +1,10 @@
+package Blackjack.exception;
+
+public class ErrorException extends IllegalArgumentException {
+
+    private static final String ERROR_PREFIX = "[ERROR] ";
+
+    public ErrorException(String message) {
+        super(ERROR_PREFIX + message);
+    }
+}
