--- src/main/java/domain/card/Card.java
@@ -1,8 +1,0 @@
-package domain.card;
-
-public record Card(Rank rank, Suit suit) {
-
-    public boolean isAce() {
-        return this.rank == Rank.ACE;
-    }
-}
