--- src/main/java/domain/card/Suit.java
@@ -1,19 +1,0 @@
-package domain.card;
-
-public enum Suit {
-
-    SPADE("스페이드"),
-    HEART("하트"),
-    DIAMOND("다이아몬드"),
-    CLOVER("클로버");
-
-    private final String name;
-
-    Suit(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
