--- src/main/java/dto/GameResultDto.java
@@ -1,7 +1,0 @@
-package dto;
-
-import domain.GameStatus;
-import java.util.Map;
-
-public record GameResultDto(String name, Map<GameStatus, Integer> gameResults) {
-}
