--- src/main/java/dto/ParticipantCardsDto.java
@@ -1,7 +1,0 @@
-package dto;
-
-import domain.card.Card;
-import java.util.List;
-
-public record ParticipantCardsDto(String name, List<Card> cards, int cardsScore) {
-}
